migrate to spring boot 3, imports, etc
