--- mock-test-generator/src/main/kotlin/org/snapmock/generator/Store.kt
@@ -1,4 +1,0 @@
-package org.snapmock.generator
-
-class Store {
-}
