+++ snapmock-core/src/main/kotlin/org/snapmock/core/InvocationStorage.kt
