--- snapmock-spring-boot-2-tests/src/main/kotlin/org/snapmock/snap/spring/data/author/AuthorEntity.kt
@@ -1,13 +1,0 @@
-package org.snapmock.snap.spring.data.author
-
-import javax.persistence.Entity
-import javax.persistence.GeneratedValue
-import javax.persistence.GenerationType
-import javax.persistence.Id
-
-@Entity
-data class AuthorEntity(
-    @Id @GeneratedValue(strategy = GenerationType.AUTO)
-    var id: Int? = null,
-    var name: String,
-)
