--- snapmock-spring-boot-2-tests/src/main/kotlin/org/snapmock/snap/spring/data/author/AuthorRepository.kt
@@ -1,5 +1,0 @@
-package org.snapmock.snap.spring.data.author
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface AuthorRepository: JpaRepository<AuthorEntity, Int> {}
