--- snapmock-spring-boot-2-tests/src/main/kotlin/org/snapmock/snap/spring/simple/hello/HelloRepository.kt
@@ -1,18 +1,0 @@
-package org.snapmock.snap.spring.simple.hello
-
-import org.springframework.stereotype.Service
-
-@Service
-class HelloRepository {
-
-    val message: String = "Hello World"
-
-    fun post(data: String): HelloData {
-        return HelloData(data)
-    }
-
-    fun getNull(_nothing: Nothing?): HelloData? {
-        return null
-    }
-
-}
