+++ snapmock-spring-boot-tests/src/main/kotlin/org/snapmock/snap/spring/data/author/AuthorRepository.kt
@@ -1,0 +1,5 @@
+package org.snapmock.snap.spring.data.author
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface AuthorRepository: JpaRepository<AuthorEntity, Int>
