+++ snapmock-spring-boot-tests/src/main/kotlin/org/snapmock/snap/spring/data/book/BookDto.kt
