+++ snapmock-spring-boot-tests/src/main/kotlin/org/snapmock/snap/spring/simple/hello/HelloData.kt
