+++ snapmock-spring-boot-tests/src/main/kotlin/org/snapmock/snap/spring/simple/hello/HelloRepository.kt
@@ -1,0 +1,16 @@
+package org.snapmock.snap.spring.simple.hello
+
+import org.springframework.stereotype.Service
+
+@Service
+class HelloRepository {
+
+    val message: String = "Hello World"
+
+    fun post(data: String): HelloData {
+        return HelloData(data)
+    }
+
+    fun getNull(nothing: Nothing?): HelloData? = null
+
+}
