+++ snapmock-spring-boot-tests/src/main/resources/data/application.properties
