+++ snapmock-spring-boot-tests/src/main/resources/simple/application.properties
